requirements.txt
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
Deprecated==1.2.18
distro==1.9.0
duckdb==1.4.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
human_mouse==0.1.2
idna==3.11
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.3
langchain-core==1.0.3
langchain-experimental==0.0.42
langchain-openai==1.0.2
langgraph==1.0.2
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.40
MouseInfo==0.1.3
mss==10.1.0
nodriver==0.47.0
numpy==2.2.6
openai==2.7.1
opencv-python==4.12.0.88
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pillow==12.0.0
PyAutoGUI==0.9.54
pydantic==2.12.3
pydantic_core==2.41.4
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python3-xlib==0.15
pytweening==1.2.0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
scipy==1.16.2
sniffio==1.3.1
soupsieve==2.8
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
wrapt==1.17.3
xxhash==3.6.0
zstandard==0.25.0
